rename component to LiveHeaderMedia
