+++ src/app/components/LiveHeaderMedia/index.stories.tsx
@@ -2,7 +2,7 @@
 /** @jsx jsx */
 import { css, jsx, Theme } from '@emotion/react';
 import mundoLiveFixture from '#data/mundo/live/c7dkx155e626t.json';
-import LiveMediaStream from '.';
+import LiveHeaderMedia from '.';
 import { MediaCollection } from '../MediaLoader/types';
 import metadata from './metadata.json';
 
@@ -22,7 +22,7 @@ export const Component = ({ warnings }: Props) => {
 
   return (
     <div css={({ palette }: Theme) => css({ background: palette.BLACK })}>
-      <LiveMediaStream mediaCollection={fixtureData as MediaCollection[]} />
+      <LiveHeaderMedia mediaCollection={fixtureData as MediaCollection[]} />
     </div>
   );
 };
